Default state for new samplesets is now unstarted
Samplesets created after analysis are built when work is reported
completed. However, these samplesets belong to a newly created batch or
even round and the victim client code is still unaware of them. They
need to be marked as unstarted and incomplete by default; when the
client later requests work, they will be marked as started as required.